python_worker/requirements.txt
python-dotenv
httpx
pika
APScheduler
